Log functionality is not organized #100
